+++ internal/checker/utils.go
@@ -303,115 +303,3 @@ func (c *Checker) validateTypeParams(
 
 	return errors
 }
-
-// extractTypeParamRefsFromTypeAnn extracts all type parameter names referenced in a type annotation
-func extractTypeParamRefsFromTypeAnn(typeAnn ast.TypeAnn, availableTypeParams map[string]bool) []string {
-	if typeAnn == nil {
-		return nil
-	}
-
-	refs := make(map[string]bool)
-	var visitor typeParamRefCollector
-	visitor.availableTypeParams = availableTypeParams
-	visitor.refs = refs
-	typeAnn.Accept(&visitor)
-
-	// Convert map to slice
-	result := make([]string, 0, len(refs))
-	for ref := range refs {
-		result = append(result, ref)
-	}
-	return result
-}
-
-// typeParamRefCollector is a visitor that collects type parameter references
-type typeParamRefCollector struct {
-	ast.DefaultVisitor
-	availableTypeParams map[string]bool
-	refs                map[string]bool
-}
-
-func (v *typeParamRefCollector) EnterTypeAnn(typeAnn ast.TypeAnn) bool {
-	if ref, ok := typeAnn.(*ast.TypeRefTypeAnn); ok {
-		// Check if this is a simple identifier (not qualified)
-		if ident, ok := ref.Name.(*ast.Ident); ok {
-			name := ident.Name
-			if v.availableTypeParams[name] {
-				v.refs[name] = true
-			}
-		}
-	}
-	return true
-}
-
-// sortTypeParamsTopologically sorts type parameters so that dependencies come before dependents
-func sortTypeParamsTopologically(typeParams []*ast.TypeParam) []*ast.TypeParam {
-	if len(typeParams) <= 1 {
-		return typeParams
-	}
-
-	// Build a map of type parameter names for quick lookup
-	typeParamNames := make(map[string]bool)
-	for _, tp := range typeParams {
-		typeParamNames[tp.Name] = true
-	}
-
-	// Build dependency graph: map from type param name to list of type params it depends on
-	deps := make(map[string][]string)
-	for _, tp := range typeParams {
-		var tpDeps []string
-		if tp.Constraint != nil {
-			tpDeps = append(tpDeps, extractTypeParamRefsFromTypeAnn(tp.Constraint, typeParamNames)...)
-		}
-		if tp.Default != nil {
-			tpDeps = append(tpDeps, extractTypeParamRefsFromTypeAnn(tp.Default, typeParamNames)...)
-		}
-		deps[tp.Name] = tpDeps
-	}
-
-	// Perform topological sort using DFS
-	sorted := make([]*ast.TypeParam, 0, len(typeParams))
-	visited := make(map[string]bool)
-	visiting := make(map[string]bool) // For cycle detection
-
-	var visit func(string) bool
-	visit = func(name string) bool {
-		if visited[name] {
-			return true // Already processed
-		}
-		if visiting[name] {
-			// Cycle detected - return false to indicate we should keep original order
-			return false
-		}
-
-		visiting[name] = true
-		for _, depName := range deps[name] {
-			if !visit(depName) {
-				return false
-			}
-		}
-		visiting[name] = false
-		visited[name] = true
-
-		// Find the type param and add it to sorted list
-		for _, tp := range typeParams {
-			if tp.Name == name {
-				sorted = append(sorted, tp)
-				break
-			}
-		}
-		return true
-	}
-
-	// Visit all type parameters
-	for _, tp := range typeParams {
-		if !visited[tp.Name] {
-			if !visit(tp.Name) {
-				// Cycle detected, return original order
-				return typeParams
-			}
-		}
-	}
-
-	return sorted
-}
